agent-python/requirements.txt
customtkinter
requests
pysnmp
python-dotenv
Pillow
